# Readability enhancements for line wrapping