--- a/shaping/wrapping.go
+++ b/shaping/wrapping.go
@@ -1,6 +1,7 @@
 package shaping
 
 import (
+	"fmt"
 	"sort"
 
 	"github.com/go-text/typesetting/di"
@@ -366,7 +367,7 @@ func (r *runMapper) mapRun(runIdx int, run Output) {
 type RunIterator interface {
 	// Next returns the next logical run index, shaped run, and true if there are more runs,
 	// then it advances the iterator to the next element.
-	// Otherwise it returns an undefined index, an empty Output, and false if there are no more outputs.
+	// Otherwise it returns an undefined index, an empty Output, and false.
 	Next() (index int, run Output, isValid bool)
 	// Peek returns the same thing Next() would, but does not advance the iterator (so the
 	// next call to Next() will return the same thing).
@@ -384,26 +385,31 @@ type RunIterator interface {
 
 // runSlice is a [RunIterator] built from already-shaped text.
 type runSlice struct {
-	runs      []Output
-	idx       int
-	savedRuns []Output
-	savedIdx  int
+	runs     []Output
+	idx      int
+	savedIdx int
 }
 
 var _ RunIterator = (*runSlice)(nil)
 
 // NewSliceIterator returns a [RunIterator] backed by an already-shaped slice of [Output]s.
-func NewSliceIterator(outs ...Output) RunIterator {
+func NewSliceIterator(outs []Output) RunIterator {
 	return &runSlice{
 		runs: outs,
 	}
+}
+
+// Reset configures the runSlice for reuse with the given shaped text.
+func (r *runSlice) Reset(outs []Output) {
+	r.runs = outs
+	r.idx = 0
+	r.savedIdx = 0
 }
 
 // Next implements [RunIterator.Next].
 func (r *runSlice) Next() (int, Output, bool) {
 	idx, run, ok := r.Peek()
 	if ok {
-		r.runs = r.runs[1:]
 		r.idx++
 	}
 	return idx, run, ok
@@ -411,29 +417,143 @@ func (r *runSlice) Next() (int, Output, bool) {
 
 // Peek implements [RunIterator.Peek].
 func (r *runSlice) Peek() (int, Output, bool) {
-	if len(r.runs) > 0 {
-		next := r.runs[0]
-		idx := r.idx
-		return idx, next, true
+	if r.idx < len(r.runs) {
+		next := r.runs[r.idx]
+		return r.idx, next, true
 	}
 	return r.idx, Output{}, false
 }
 
 // IsFinal implements [RunIterator.IsFinal].
 func (r *runSlice) IsFinal() bool {
-	return len(r.runs) == 1
+	return r.idx == len(r.runs)-1
 }
 
 // Save implements [RunIterator.Save].
 func (r *runSlice) Save() {
 	r.savedIdx = r.idx
-	r.savedRuns = r.runs
 }
 
 // Restore implements [RunIterator.Restore].
 func (r *runSlice) Restore() {
 	r.idx = r.savedIdx
-	r.runs = r.savedRuns
+}
+
+// WrapBuffer provides reusable buffers for line wrapping. When using a
+// WrapBuffer, returned line wrapping results will use memory stored within
+// the buffer. This means that the same buffer cannot be reused for another
+// wrapping operation while the wrapped lines are still in use (unless they
+// are deeply copied). If necessary, using a multiple WrapBuffers can work
+// around this restriction.
+type WrapBuffer struct {
+	// paragraph is a buffer holding paragraph allocated (primarily) from subregions
+	// of the line field.
+	paragraph []Line
+	// line is a large buffer of Outputs that is used to build lines.
+	line []Output
+	// lineUsed is the index of the first unused element in line.
+	lineUsed int
+	// lineExhausted indicates whether the previous shaping used all of line.
+	lineExhausted bool
+	// alt is a smaller temporary buffer that holds candidate lines while
+	// they are being built.
+	alt []Output
+	// best is a slice holding the best known line. When possible, it
+	// is a subslice of line, but if line runs out of capacity it will
+	// be heap allocated.
+	best []Output
+	// bestInLine tracks whether the current best is allocated from within line.
+	bestInLine bool
+}
+
+// NewWrapBuffer returns a WrapBuffer with some pre-allocated storage for
+// small-sized texts.
+func NewWrapBuffer() *WrapBuffer {
+	return &WrapBuffer{
+		paragraph: make([]Line, 0, 10),
+		line:      make([]Output, 0, 100),
+		alt:       make([]Output, 0, 10),
+	}
+}
+
+func (w *WrapBuffer) reset() {
+	w.paragraph = w.paragraph[:0]
+	w.alt = w.alt[:0]
+	w.line = w.line[:0]
+	w.lineUsed = 0
+	w.best = nil
+	w.bestInLine = false
+	if w.lineExhausted {
+		w.lineExhausted = false
+		// Trigger the go slice growth heuristic by appending an element to
+		// the capacity.
+		w.line = append(w.line[:cap(w.line)], Output{})[:0]
+	}
+}
+
+func (w *WrapBuffer) stats() string {
+	return fmt.Sprintf("paragraph: %d(%d), line: %d(%d), used: %d, exhausted: %v, alt: %d(%d)", len(w.paragraph), cap(w.paragraph), len(w.line), cap(w.line), w.lineUsed, w.lineExhausted, len(w.alt), cap(w.alt))
+}
+
+// singleRunParagraph is an optimized helper for quickly constructing
+// a []Line containing only a single run.
+func (w *WrapBuffer) singleRunParagraph(run Output) []Line {
+	w.paragraph = w.paragraph[:0]
+	s := w.line[w.lineUsed : w.lineUsed+1]
+	s[0] = run
+	w.paragraphAppend(s)
+	return w.finalParagraph()
+}
+
+func (w *WrapBuffer) paragraphAppend(line []Output) {
+	w.paragraph = append(w.paragraph, line)
+}
+
+func (w *WrapBuffer) finalParagraph() []Line {
+	return w.paragraph
+}
+
+func (w *WrapBuffer) startLine() {
+	w.alt = w.alt[:0]
+	w.best = nil
+	w.bestInLine = false
+}
+
+// candidateAppend adds the given run to the current line wrapping candidate.
+func (w *WrapBuffer) candidateAppend(run Output) {
+	w.alt = append(w.alt, run)
+}
+
+// markCandidateBest marks that the current line wrapping candidate is the best
+// known line wrapping candidate with the given suffixes. Providing suffixes does
+// not modify the current candidate, but does ensure that the "best" candidate ends
+// with them.
+func (w *WrapBuffer) markCandidateBest(suffixes ...Output) {
+	neededLen := len(w.alt) + len(suffixes)
+	if len(w.line[w.lineUsed:cap(w.line)]) < neededLen {
+		w.lineExhausted = true
+		w.best = make([]Output, neededLen)
+		w.bestInLine = false
+	} else {
+		w.best = w.line[w.lineUsed : w.lineUsed+neededLen]
+		w.bestInLine = true
+	}
+	n := copy(w.best, w.alt)
+	copy(w.best[n:], suffixes)
+}
+
+// hasBest returns whether there is currently a known valid line wrapping candidate
+// for the line.
+func (w *WrapBuffer) hasBest() bool {
+	return len(w.best) > 0
+}
+
+// finalizeBest commits the storage for the current best line and returns it.
+func (w *WrapBuffer) finalizeBest() []Output {
+	if w.bestInLine {
+		w.lineUsed += len(w.best)
+	}
+	return w.best
 }
 
 // LineWrapper holds reusable state for a line wrapping operation. Reusing
@@ -448,6 +568,9 @@ type LineWrapper struct {
 
 	// breaker provides line-breaking candidates.
 	breaker *breaker
+
+	// scratch holds wrapping algorithm storage buffers for reuse.
+	scratch *WrapBuffer
 
 	// mapper tracks rune->glyphCluster mappings.
 	mapper runMapper
@@ -467,8 +590,11 @@ type LineWrapper struct {
 }
 
 // Prepare initializes the LineWrapper for the given paragraph and shaped text.
-// It must be called prior to invoking WrapNextLine.
-func (l *LineWrapper) Prepare(config WrapConfig, paragraph []rune, runs RunIterator) {
+// It must be called prior to invoking WrapNextLine. The scratch parameter can
+// be used to avoid allocations by allocating wrapped lines from within the buffer.
+// If scratch is non-nil, lines returned by the wrapper are only valid until
+// buffer is reused in a line wrapping operation. See [WrapBuffer] for details.
+func (l *LineWrapper) Prepare(config WrapConfig, paragraph []rune, runs RunIterator, scratch *WrapBuffer) {
 	l.config = config
 	l.truncating = l.config.TruncateAfterLines > 0
 	l.breaker = newBreaker(&l.seg, paragraph)
@@ -477,26 +603,39 @@ func (l *LineWrapper) Prepare(config WrapConfig, paragraph []rune, runs RunItera
 	l.lineStartRune = 0
 	l.more = true
 	l.mapper.valid = false
+	l.scratch = scratch
+	if l.scratch == nil {
+		l.scratch = NewWrapBuffer()
+	}
+	l.scratch.reset()
 }
 
 // WrapParagraph wraps the paragraph's shaped glyphs to a constant maxWidth.
 // It is equivalent to iteratively invoking WrapLine with a constant maxWidth.
 // If the config has a non-zero TruncateAfterLines, WrapParagraph will return at most
 // that many lines. The truncated return value is the count of runes truncated from
-// the end of the text.
-func (l *LineWrapper) WrapParagraph(config WrapConfig, maxWidth int, paragraph []rune, runs RunIterator) (_ []Line, truncated int) {
-	if _, nextRun, ok := runs.Peek(); ok && runs.IsFinal() && nextRun.Advance.Ceil() < maxWidth && !(config.TextContinues && config.TruncateAfterLines == 1) {
-		return []Line{[]Output{nextRun}}, 0
+// the end of the text. If non-nil, the provided scratch buffer will be used
+// to store as much of the returned lines as possible. This eliminates allocations,
+// but means that the buffer cannot be reused while the returned lines are in
+// use. Applications must either deep copy or finish processing those lines
+// before reusing the same buffer. See [WrapBuffer] for details.
+func (l *LineWrapper) WrapParagraph(config WrapConfig, maxWidth int, paragraph []rune, runs RunIterator, scratch *WrapBuffer) (_ []Line, truncated int) {
+	if scratch == nil {
+		scratch = NewWrapBuffer()
+	} else {
+		scratch.reset()
 	}
-	l.Prepare(config, paragraph, runs)
-	var lines []Line
+	if _, nextRun, ok := runs.Peek(); ok && runs.IsFinal() && nextRun.Advance.Ceil() < maxWidth && !(config.TextContinues && config.TruncateAfterLines == 1) {
+		return scratch.singleRunParagraph(nextRun), 0
+	}
+	l.Prepare(config, paragraph, runs, scratch)
 	var done bool
 	for !done {
 		var line Line
 		line, truncated, done = l.WrapNextLine(maxWidth)
-		lines = append(lines, line)
+		scratch.paragraphAppend(line)
 	}
-	return lines, truncated
+	return scratch.finalParagraph(), truncated
 }
 
 // nextBreakOption returns the next rune offset at which the line can be broken,
@@ -533,13 +672,13 @@ const (
 // fillUntil tries to fill the provided line candidate slice with runs until it reaches a run containing the
 // provided break option. It returns the index of the run containing the option, the new width of the candidate
 // line, the contents of the new candidate line, and a result indicating how to proceed.
-func (l *LineWrapper) fillUntil(runs RunIterator, option breakOption, startWidth fixed.Int26_6, lineCandidate []Output) (newWidth fixed.Int26_6, newLineCandidate []Output, status fillResult) {
+func (l *LineWrapper) fillUntil(runs RunIterator, option breakOption, startWidth fixed.Int26_6, scratch *WrapBuffer) (newWidth fixed.Int26_6, status fillResult) {
 	currRunIndex, run, more := runs.Peek()
 	for more && option.breakAtRune >= run.Runes.Count+run.Runes.Offset {
 		if l.lineStartRune >= run.Runes.Offset+run.Runes.Count {
 			_, _, isMore := runs.Next()
 			if !isMore {
-				return startWidth, lineCandidate, noCandidate
+				return startWidth, noCandidate
 			}
 			currRunIndex, run, more = runs.Peek()
 			continue
@@ -551,15 +690,15 @@ func (l *LineWrapper) fillUntil(runs RunIterator, option breakOption, startWidth
 		}
 		// While the run being processed doesn't contain the current line breaking
 		// candidate, just append it to the candidate line.
-		lineCandidate = append(lineCandidate, run)
+		scratch.candidateAppend(run)
 		startWidth += run.Advance
 		_, _, isMore := runs.Next()
 		if !isMore {
-			return startWidth, lineCandidate, noRunWithBreak
+			return startWidth, noRunWithBreak
 		}
 		currRunIndex, run, more = runs.Peek()
 	}
-	return startWidth, lineCandidate, newCandidate
+	return startWidth, newCandidate
 }
 
 // WrapNextLine wraps the shaped glyphs of a paragraph to a particular max width.
@@ -569,6 +708,7 @@ func (l *LineWrapper) fillUntil(runs RunIterator, option breakOption, startWidth
 // The truncated return value is the count of runes truncated from the end of the line,
 // if this line was truncated.
 func (l *LineWrapper) WrapNextLine(maxWidth int) (finalLine Line, truncated int, done bool) {
+	l.scratch.startLine()
 	defer func() {
 		if len(finalLine) > 0 {
 			finalRun := finalLine[len(finalLine)-1]
@@ -595,17 +735,17 @@ func (l *LineWrapper) WrapNextLine(maxWidth int) (finalLine Line, truncated int,
 		return nil, truncated, true
 	} else if _, nextRun, more := l.glyphRuns.Peek(); !more {
 		return nil, truncated, true
-	} else if len(nextRun.Glyphs) == 0 {
-		// Pass empty lines through as empty.
-		nextRun.Runes = Range{Count: l.breaker.totalRunes}
-		return Line([]Output{nextRun}), truncated, true
-	} else if l.glyphRuns.IsFinal() && nextRun.Advance.Ceil() < maxWidth && !(l.config.TextContinues && l.config.TruncateAfterLines == 1) {
-		return Line([]Output{nextRun}), truncated, true
+	} else if emptyLine := len(nextRun.Glyphs) == 0; emptyLine ||
+		(l.glyphRuns.IsFinal() && nextRun.Advance.Ceil() < maxWidth && !(l.config.TextContinues && l.config.TruncateAfterLines == 1)) {
+		if emptyLine {
+			// Pass empty lines through as empty.
+			nextRun.Runes = Range{Count: l.breaker.totalRunes}
+		}
+		l.scratch.candidateAppend(nextRun)
+		l.scratch.markCandidateBest()
+		return l.scratch.finalizeBest(), truncated, true
 	}
 
-	// lineCandidate is filled with runs as we search for valid line breaks. When we find a valid
-	// option, we commit it into bestCandidate and keep looking.
-	var lineCandidate, bestCandidate []Output
 	// lineWidth tracks the width of the lineCandidate.
 	lineWidth := fixed.I(0)
 	var result fillResult
@@ -623,19 +763,20 @@ func (l *LineWrapper) WrapNextLine(maxWidth int) (finalLine Line, truncated int,
 
 		option, ok := l.nextBreakOption()
 		if !ok {
-			return bestCandidate, truncated, true
+			return l.scratch.finalizeBest(), truncated, true
 		}
-		lineWidth, lineCandidate, result = l.fillUntil(
+		lineWidth, result = l.fillUntil(
 			l.glyphRuns,
 			option,
 			lineWidth,
-			lineCandidate,
+			l.scratch,
 		)
 
 		if result == noCandidate {
-			return bestCandidate, truncated, true
+			return l.scratch.finalizeBest(), truncated, true
 		} else if result == noRunWithBreak {
-			return lineCandidate, truncated, true
+			l.scratch.markCandidateBest()
+			return l.scratch.finalizeBest(), truncated, true
 		}
 		currRunIndex, run, _ := l.glyphRuns.Peek()
 		l.mapper.mapRun(currRunIndex, run)
@@ -647,22 +788,22 @@ func (l *LineWrapper) WrapNextLine(maxWidth int) (finalLine Line, truncated int,
 		candidateLineWidth := (candidateRun.Advance + lineWidth).Ceil()
 		if candidateLineWidth > maxWidth {
 			// The run doesn't fit on the line.
-			if len(bestCandidate) < 1 {
+			if !l.scratch.hasBest() {
 				if truncating {
-					return bestCandidate, truncated, true
+					return l.scratch.finalizeBest(), truncated, true
 				}
 				// There is no existing candidate that fits, and we have just hit the
 				// first line breaking candidate. Commit this break position as the
 				// best available, even though it doesn't fit.
-				lineCandidate = append(lineCandidate, candidateRun)
-				return lineCandidate, truncated, false
+				l.scratch.markCandidateBest(candidateRun)
+				return l.scratch.finalizeBest(), truncated, false
 			} else {
 				// The line is a valid, shorter wrapping. Return it and mark that
 				// we should reuse the current line break candidate on the next
 				// line.
 				l.isUnused = true
 				l.glyphRuns.Restore()
-				return bestCandidate, truncated, false
+				return l.scratch.finalizeBest(), truncated, false
 			}
 		} else if truncating && candidateLineWidth > truncatedMaxWidth {
 			// The run would not fit if truncated.
@@ -670,46 +811,17 @@ func (l *LineWrapper) WrapNextLine(maxWidth int) (finalLine Line, truncated int,
 			if finalRunRune == l.breaker.totalRunes && !l.config.TextContinues {
 				// The run contains the entire end of the text, so no truncation is
 				// necessary.
-				bestCandidate = commitCandidate(bestCandidate, lineCandidate, candidateRun)
-				return bestCandidate, truncated, true
+				l.scratch.markCandidateBest(candidateRun)
+				return l.scratch.finalizeBest(), truncated, true
 			}
 			// We must truncate the line in order to show it.
-			return bestCandidate, truncated, true
+			return l.scratch.finalizeBest(), truncated, true
 		} else {
 			// The run does fit on the line. Commit this line as the best known
 			// line, but keep lineCandidate unmodified so that later break
 			// options can be attempted to see if a more optimal solution is
 			// available.
-			bestCandidate = commitCandidate(bestCandidate, lineCandidate, candidateRun)
+			l.scratch.markCandidateBest(candidateRun)
 		}
 	}
-}
-
-// commitCandidate efficiently updates destination to contain append(source, newRuns...),
-// returning the resulting slice. This operation only makes sense when destination
-// is not known to contain the elements of source already.
-func commitCandidate(destination, source []Output, newRuns ...Output) []Output {
-	destination = resize(destination, len(source), len(source)+1)
-	destination = destination[:copy(destination, source)]
-	destination = append(destination, newRuns...)
-	return destination
-}
-
-// resize returns input resized to have the provided length and at least the provided
-// capacity. It may copy the data if the provided capacity is greater than the capacity
-// of in. If the provided length is greater than the provided capacity, the capacity will
-// be used as the length.
-func resize(input []Output, length, capacity int) []Output {
-	if length > capacity {
-		length = capacity
-	}
-	out := input
-	if cap(input) < capacity {
-		out = make([]Output, capacity)
-		copy(out, input)
-	}
-	if len(out) != length {
-		out = out[:length]
-	}
-	return out
 }