@whereswaldon 2023-05-16
I've changed my mind about splitting this changeset from #68, so I've incorporated these changes there.